tests(Logging): Use same logger in CLI and tests
