+++ src/cli.ts
@@ -32,7 +32,7 @@
  * Please see the documentation for each function on the arguments required and
  * options available.
  */
-import * as logga from '@stencila/logga'
+import {default as log, configure} from './log'
 import minimist from 'minimist'
 import path from 'path'
 import { convert, read, write } from '.'
@@ -53,23 +53,9 @@ const { _, ...options } = minimist(process.argv.slice(2), {
 const command = _[0]
 const args = _.slice(1)
 
-/**
- * Set up logger so that it:
- *
- * - only shows DEBUG entries if --debug=true
- * - does not show duplicate entries unless --debug=true
- */
-const log = logga.getLogger('encoda:cli')
-const previousLogData = new Set<string>()
-logga.replaceHandlers((data: logga.LogData): void => {
-  if (data.level <= (options.debug ? 3 : 2)) {
-    const json = JSON.stringify(data)
-    if (options.debug || !previousLogData.has(json)) {
-      logga.defaultHandler(data)
-      previousLogData.add(json)
-    }
-  }
-})
+// Configure the log
+configure(options.debug)
+
 //  eslint-disable-next-line
 ;(async () => {
   try {
